refactor: bingads uploader into multiple files
